% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{get_saxonian}
\alias{get_saxonian}
\title{get_saxonion}
\usage{
get_saxonian(txt)
}
\arguments{
\item{txt}{provide a text you want to translate from German into Saxonian}
}
\description{
Translates German to Saxonian
}
